pyproject.toml
[project]
name = "temporal-context-mcp"
version = "0.1.0"
description = "A MCP server that provides intelligent temporal context to adapt responses according to the user's time and routines."
readme = "README.md"
authors = [
    { name = "Federico Gomez", email = "gomez00federico@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "croniter>=6.0.0",
    "mcp[cli]>=1.12.0",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0.post0",
]

[project.scripts]
temporal-context-mcp = "temporal_context_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
